Add transfer from and to account

+++ db/migrate/20160201145236_change_precision_and_default_values_to_money_fields.rb
@@ -1,6 +1,6 @@
 class ChangePrecisionAndDefaultValuesToMoneyFields < ActiveRecord::Migration
   def change
-    change_column :accounts, :funds, :integer, { precision: 10, scale: 2, default: 0.00 }
-    change_column :transactions, :amount, :integer, { precision:10, scale: 2, default: 0.00 }
+    change_column :accounts, :funds, :decimal, { precision: 10, scale: 2, default: 0.00 }
+    change_column :transactions, :amount, :decimal, { precision:10, scale: 2, default: 0.00 }
   end
 end
